cloud-run-document-processor/requirements.txt
flask
google-cloud-documentai
google-cloud-firestore
google-cloud-pubsub
